Rename and expose internal modules
This also removes some "Stability" annotations from internal modules.

Context: #211.
