+++ Data/HashMap/Internal/List.hs
@@ -1,9 +1,24 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# OPTIONS_GHC -fno-full-laziness -funbox-strict-fields #-}
--- | Extra list functions
+
+-- | = WARNING
+--
+-- This module is considered __internal__.
+--
+-- The Package Versioning Policy __does not apply__.
+--
+-- The contents of this module may change __in any way whatsoever__
+-- and __without any warning__ between minor versions of this package.
+--
+-- Authors importing this module are expected to track development
+-- closely.
+--
+-- = Description
+--
+-- Extra list functions
 --
 -- In separate module to aid testing.
-module Data.HashMap.List
+module Data.HashMap.Internal.List
     ( isPermutationBy
     , deleteBy
     , unorderedCompare
